+++ packages/clean_framework_router/CHANGELOG.md
@@ -1,4 +1,15 @@
 # Changelog
+## 0.2.0-dev.2
+**Nov 3, 2022**
+- Added `AppRouterState` alias.
+
+## 0.2.0-dev.1
+**Nov 2, 2022 (Breaking)**
+- Upgraded to `go_router` v5.
+- `AppRouter` is now abstract, in order to make it extensible.
+- Routes now need to be enum mixed with `RoutesMixin`.
+- Added `AppRouterScope`.
+
 ## 0.2.0-dev.0
 **Nov 1, 2022**
 - Upgraded to `clean_framework: ^2.0.0-dev.0`.